move declarative config types
