+++ crates/goose/src/config/declarative_providers.rs
@@ -10,21 +10,15 @@ use crate::providers::openai_def::OpenAiProviderDef;
 use anyhow::Result;
 use include_dir::{include_dir, Dir};
 use once_cell::sync::Lazy;
-use serde::{Deserialize, Deserializer, Serialize};
+use serde::{Deserialize, Serialize};
 use std::str::FromStr;
 
-/// Deserialize an optional string, treating empty/whitespace-only values as None.
-fn deserialize_non_empty_string<'de, D>(deserializer: D) -> Result<Option<String>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let opt: Option<String> = Option::deserialize(deserializer)?;
-    Ok(opt.filter(|s| !s.trim().is_empty()))
-}
 use std::collections::HashMap;
 use std::path::{Path, PathBuf};
 use std::sync::Mutex;
 use utoipa::ToSchema;
+
+pub use goose_providers::declarative::*;
 
 static FIXED_PROVIDERS: Dir = include_dir!("$CARGO_MANIFEST_DIR/src/providers/declarative");
 
@@ -32,102 +26,8 @@ pub fn custom_providers_dir() -> std::path::PathBuf {
     Paths::config_dir().join("custom_providers")
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, ToSchema)]
-#[serde(rename_all = "lowercase")]
-pub enum ProviderEngine {
-    OpenAI,
-    Ollama,
-    Anthropic,
-}
-
-impl FromStr for ProviderEngine {
-    type Err = anyhow::Error;
-
-    fn from_str(engine: &str) -> Result<Self> {
-        match engine.trim().to_lowercase().as_str() {
-            "openai" | "openai_compatible" => Ok(Self::OpenAI),
-            "anthropic" | "anthropic_compatible" => Ok(Self::Anthropic),
-            "ollama" | "ollama_compatible" => Ok(Self::Ollama),
-            _ => Err(anyhow::anyhow!("Invalid provider type: {}", engine)),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, ToSchema)]
-pub struct EnvVarConfig {
-    pub name: String,
-    #[serde(default)]
-    pub required: bool,
-    #[serde(default)]
-    pub secret: bool,
-    /// When true, the field is shown prominently in the UI (not collapsed).
-    /// Defaults to the value of `required` if not specified.
-    pub primary: Option<bool>,
-    pub description: Option<String>,
-    pub default: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, ToSchema)]
-pub struct DeclarativeProviderConfig {
-    pub name: String,
-    pub engine: ProviderEngine,
-    pub display_name: String,
-    pub description: Option<String>,
-    #[serde(default)]
-    pub api_key_env: String,
-    pub base_url: String,
-    pub models: Vec<ModelInfo>,
-    pub headers: Option<HashMap<String, String>>,
-    pub timeout_seconds: Option<u64>,
-    pub supports_streaming: Option<bool>,
-    #[serde(default = "default_requires_auth")]
-    pub requires_auth: bool,
-    #[serde(default)]
-    pub catalog_provider_id: Option<String>,
-    #[serde(default)]
-    pub base_path: Option<String>,
-    #[serde(default)]
-    pub env_vars: Option<Vec<EnvVarConfig>>,
-    /// Controls whether `fetch_supported_models` calls the provider's `/v1/models`
-    /// endpoint or returns the static `models` list directly.
-    ///
-    /// - `Some(false)` + non-empty `models`: return the static list; no API call.
-    ///   Construction fails if `models` is empty.
-    /// - `Some(true)` or `None`: try the API; fall back to `models` on 404.
-    #[serde(default)]
-    pub dynamic_models: Option<bool>,
-    #[serde(default)]
-    pub skip_canonical_filtering: bool,
-    #[serde(default, deserialize_with = "deserialize_non_empty_string")]
-    pub model_doc_link: Option<String>,
-    #[serde(default)]
-    pub setup_steps: Vec<String>,
-    #[serde(default, deserialize_with = "deserialize_non_empty_string")]
-    pub fast_model: Option<String>,
-    #[serde(default)]
-    pub preserves_thinking: bool,
-}
-
-fn default_requires_auth() -> bool {
-    true
-}
-
 fn should_preserve_thinking_by_default(engine: &ProviderEngine) -> bool {
     matches!(engine, ProviderEngine::OpenAI)
-}
-
-impl DeclarativeProviderConfig {
-    pub fn id(&self) -> &str {
-        &self.name
-    }
-
-    pub fn display_name(&self) -> &str {
-        &self.display_name
-    }
-
-    pub fn models(&self) -> &[ModelInfo] {
-        &self.models
-    }
 }
 
 /// Expand `${VAR_NAME}` placeholders in a template string using the given env var configs.
